backend: checkstyle fixes

+++ backend/src/test/java/de/learnlib/alex/data/dao/SymbolDAOTest.java
@@ -16,6 +16,13 @@
 
 package de.learnlib.alex.data.dao;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.verify;
+
 import com.fasterxml.jackson.databind.ObjectMapper;
 import de.learnlib.alex.auth.entities.User;
 import de.learnlib.alex.common.exceptions.NotFoundException;
@@ -56,12 +63,6 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.transaction.TransactionSystemException;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
-import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.doThrow;
-import static org.mockito.Mockito.verify;
 
 @RunWith(MockitoJUnitRunner.class)
 public class SymbolDAOTest {
@@ -578,7 +579,7 @@ public class SymbolDAOTest {
 
         for (int id = 0; id < symbols.size(); id++) {
             Symbol symbol = symbols.get(id);
-            long symbolId = (long) id;
+            long symbolId = id;
             symbol.setId(symbolId);
         }
